ormConvertTupleNew() bugfix for PHP 8+
method_exists(), starting in PHP 8.0 no longer accepts the first argument to be anything other than a string or an object. Before, it would sometimes be passed in method_exists(null, "onBeforeLoad"), which would be valid. Of course, it would return false. However, now, we must perform an external check to see if the first argument is null, and then never execute method_exists()